--- src/components/Slider/__stories__/examples/SliderExampleBasic/SliderExampleBasic.tsx
@@ -1,19 +1,0 @@
-import React, { useState } from 'react';
-
-import { StoryBookExample } from '../../../../../uiKit/components/StoryBookExample/StoryBookExample';
-import { Slider } from '../../../Slider';
-
-export const SliderExampleBasic = () => {
-  const [value, setValue] = useState<number>(20);
-
-  return (
-    <StoryBookExample>
-      <Slider
-        range={false}
-        label={`Значение ${value}`}
-        onChange={({ value }) => setValue(value)}
-        value={value}
-      />
-    </StoryBookExample>
-  );
-};
