--- src/components/Slider/__stories__/examples/SliderExampleSize/SliderExampleSize.tsx
@@ -1,45 +1,0 @@
-import './SliderExampleSize.css';
-
-import React, { useState } from 'react';
-
-import { cn } from '../../../../../utils/bem';
-import { Slider } from '../../../Slider';
-
-const cnSliderExampleSize = cn('SliderExampleSize');
-
-export const SliderExampleSize = () => {
-  const [value, setValue] = useState<number>(20);
-
-  return (
-    <div className={cnSliderExampleSize()}>
-      <Slider
-        range={false}
-        label={`Значение ${value}`}
-        onChange={({ value }) => setValue(value)}
-        value={value}
-        size="xs"
-      />
-      <Slider
-        range={false}
-        label={`Значение ${value}`}
-        onChange={({ value }) => setValue(value)}
-        value={value}
-        size="s"
-      />
-      <Slider
-        range={false}
-        label={`Значение ${value}`}
-        onChange={({ value }) => setValue(value)}
-        value={value}
-        size="m"
-      />
-      <Slider
-        range={false}
-        label={`Значение ${value}`}
-        onChange={({ value }) => setValue(value)}
-        value={value}
-        size="l"
-      />
-    </div>
-  );
-};
