+++ src/components/SliderCanary/Slider.css
