+++ src/components/SliderCanary/SliderLine/SliderLine.tsx
