+++ src/components/SliderCanary/SliderPoint/SliderPoint.css
