+++ src/components/SliderCanary/__stories__/examples/SliderExampleSize/SliderExampleSize.css
