+++ src/components/SliderCanary/__stories__/examples/SliderExampleStatus/SliderExampleStatus.tsx
@@ -1,0 +1,48 @@
+import './SliderExampleStatus.css';
+
+import React, { useState } from 'react';
+
+import { cn } from '../../../../../utils/bem';
+import { Slider } from '../../../SliderCanary';
+
+const cnSliderExampleStatus = cn('SliderExampleStatus');
+
+export const SliderExampleStatus = () => {
+  const [value, setValue] = useState<number>(20);
+
+  return (
+    <div className={cnSliderExampleStatus()}>
+      <Slider
+        range={false}
+        label={`Значение ${value}`}
+        caption="Статус alert"
+        onChange={({ value }) => setValue(value)}
+        value={value}
+        status="alert"
+      />
+      <Slider
+        range={false}
+        label={`Значение ${value}`}
+        caption="Статус warning"
+        onChange={({ value }) => setValue(value)}
+        value={value}
+        status="warning"
+      />
+      <Slider
+        range={false}
+        label={`Значение ${value}`}
+        caption="Статус success"
+        onChange={({ value }) => setValue(value)}
+        value={value}
+        status="success"
+      />
+      <Slider
+        range={false}
+        label={`Значение ${value}`}
+        caption="Обычный ползунок"
+        onChange={({ value }) => setValue(value)}
+        value={value}
+      />
+    </div>
+  );
+};
