+++ src/components/SliderCanary/__stories__/examples/SliderExampleStep/SliderExampleStep.tsx
@@ -1,0 +1,38 @@
+import React, { useState } from 'react';
+
+import { cnDocsDecorator } from '../../../../../uiKit/components/DocsDecorator/DocsDecorator';
+import { Slider } from '../../../SliderCanary';
+
+export const SliderExampleStepSingle = () => {
+  const [value, setValue] = useState<number>(20);
+
+  return (
+    <div className={cnDocsDecorator('Section')}>
+      <Slider
+        view="division"
+        step={10}
+        label={`Значение ${value}`}
+        onChange={({ value }) => setValue(value)}
+        value={value}
+        range={false}
+      />
+    </div>
+  );
+};
+
+export const SliderExampleStepArray = () => {
+  const [value, setValue] = useState<number>(20);
+
+  return (
+    <div className={cnDocsDecorator('Section')}>
+      <Slider
+        view="division"
+        step={[15, 40, 70]}
+        label={`Значение ${value}`}
+        onChange={({ value }) => setValue(value)}
+        value={value}
+        range={false}
+      />
+    </div>
+  );
+};
