+++ src/components/SliderCanary/__stories__/examples/SliderExampleView/SliderExampleView.tsx
@@ -1,0 +1,37 @@
+import React, { useState } from 'react';
+
+import { cnDocsDecorator } from '../../../../../uiKit/components/DocsDecorator/DocsDecorator';
+import { Slider } from '../../../SliderCanary';
+
+export const SliderExampleViewDefault = () => {
+  const [value, setValue] = useState<number>(20);
+
+  return (
+    <div className={cnDocsDecorator('Section')}>
+      <Slider
+        view="default"
+        label={`Значение ${value}`}
+        onChange={({ value }) => setValue(value)}
+        value={value}
+        range={false}
+      />
+    </div>
+  );
+};
+
+export const SliderExampleViewDivision = () => {
+  const [value, setValue] = useState<number>(20);
+
+  return (
+    <div className={cnDocsDecorator('Section')}>
+      <Slider
+        view="division"
+        step={10}
+        range={false}
+        label={`Значение ${value}`}
+        onChange={({ value }) => setValue(value)}
+        value={value}
+      />
+    </div>
+  );
+};
