+++ src/components/SliderCanary/useSlider/helper.ts
