+++ src/components/SliderCanary/useSlider/useSlider.ts
