+++ src/components/SliderCanary/useSliderStationing.ts
